refactor: organize providers and augmenters into folders
