+++ src/augmenters/mod.rs
@@ -1,0 +1,5 @@
+pub mod augmenters;
+pub mod ldap_augmenter;
+pub mod plain_augmenter;
+
+pub use augmenters::*;
