--- src/auth/mod.rs
@@ -1,10 +1,0 @@
-pub mod auth;
-pub mod ecmwfapi_provider;
-pub mod jwt_provider;
-pub mod ldap_augmenter;
-pub mod openid_offline_provider;
-pub mod plain_augmenter;
-pub mod plain_provider;
-
-// Re-export from auth.rs so we can do "use crate::auth::*;"
-pub use auth::Auth;
