+++ src/providers/mod.rs
@@ -1,0 +1,8 @@
+pub mod ecmwfapi_provider;
+pub mod jwt_provider;
+pub mod openid_offline_provider;
+pub mod plain_provider;
+pub mod providers;
+
+// Re-export from providers.rs so we can do "use crate::providers::*;"
+pub use providers::*;
